+++ CHANGELOG.md
@@ -16,6 +16,8 @@
   - The `numericRange` option removed from `Documentation` class constructor argument;
   - The `brandHandling` should consist of postprocessing functions altering the depiction made by Zod 4;
   - The `Depicter` type changed to `Overrider` having different signature;
+- The `optionalPropStyle` option removed from `Integration` class constructor:
+  - Use the new `z.interface()` schema to describe key-optional objects: https://v4.zod.dev/v4#zinterface.
 
 ## Version 23